[refactor] 마이페이지 hook 함수 수정
